Fix test/Interpreter/layout_reabstraction harder.
Swift SVN r22087

+++ test/Interpreter/layout_reabstraction.swift
@@ -75,9 +75,9 @@ println(sizeof(ContainsTupleOfFunctions<()>.self) / sizeof(Int.self))
 // CHECK-NEXT: 2
 println(unspecializedSizeOf(ContainsTupleOfFunctions<()>.self) / sizeof(Int.self))
 // CHECK-NEXT: 3
-println(sizeof(ContainsTupleOfFunctions<Int64>.self) / sizeof(Int.self))
+println(sizeof(ContainsTupleOfFunctions<Int>.self) / sizeof(Int.self))
 // CHECK-NEXT: 3
-println(unspecializedSizeOf(ContainsTupleOfFunctions<Int64>.self) / sizeof(Int.self))
+println(unspecializedSizeOf(ContainsTupleOfFunctions<Int>.self) / sizeof(Int.self))
 
 let x = ContainsTupleOfFunctions(x: (1, { $0 + 1 }))
 let y = ContainsTupleOfFunctions(x: ("foo", { $0 + "bar" }))
